frontend validations for forms

+++ db/seeds.rb
@@ -5,9 +5,43 @@
 #
 #   movies = Movie.create([{ name: "Star Wars" }, { name: "Lord of the Rings" }])
 #   Character.create(name: "Luke", movie: movies.first)
+return unless Rails.env.development?
 
+p "#{User.count} users present before delete_all"
+p "#{Inbox.count} inboxes present before delete_all"
+p "#{Message.count} messages present before delete_all"
+
+Message.delete_all
 Inbox.delete_all
-5.times do
-    faker_name = Faker::Quote.famous_last_words
-    Inbox.create(name: faker_name)
-end
+User.delete_all
+
+p "#{User.count} users present after delete_all"
+p "#{Inbox.count} inboxes present after delete_all"
+p "#{Message.count} messages present after delete_all"
+
+2.times do
+  faker_email = Faker::Internet.email
+  user = User.create(email: faker_email, 
+              password: Devise.friendly_token[0, 20])
+
+    rand(1..5).times do
+      faker_name = Faker::Quote.famous_last_words
+      inbox = Inbox.create(name: faker_name, 
+                           user: user)
+
+      rand(1..5).times do
+        message_body = Faker::Lorem.paragraph
+        Message.create(body: message_body, 
+                       inbox: inbox,
+                       user: User.first) 
+                       #user_id: rand(1..2),
+      end
+    end
+  end
+
+
+  p "#{User.count} users created"
+  p "#{Inbox.count} inboxes created"
+  p "#{Message.count} messages created"
+
+
